desloppify: move NotificationThrottler to dedicated notification.rs module
